requirements.txt
playwright==1.48.0
anthropic==0.40.0
httpx==0.27.2
python-dotenv==1.0.1
pillow==10.4.0
beautifulsoup4==4.12.3
lxml==5.3.0
colorama==0.4.6
